ProtoEmail input, Zeroize impl and SecretString, replace inquire with rpassword for safer password input and README update about security.
