+++ Cargo.toml
@@ -26,8 +26,10 @@ lettre = "0.11"
 log = "0.4.27"
 pretty_assertions = "1.4.1"
 rand = "0.9.1"
+rpassword = "7.4.0"
 ron = "0.10.1"
 rust_decimal = { version = "1.37.2", features = ["macros"] }
+secrecy = "0.10.3"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.114"
 sha2 = "0.10.9"
