+++ crates/cli/Cargo.toml
@@ -22,7 +22,9 @@ fern = "0.7.1"
 inquire = { version = "0.7.5", features = ["date"] }
 log.workspace = true
 rand.workspace = true
+rpassword.workspace = true
 ron.workspace = true
+secrecy.workspace = true
 thiserror.workspace = true
 typed-builder.workspace = true
