+++ crates/core/src/logic/encryption/encrypted_app_password.rs
@@ -1,11 +1,22 @@
 use crate::prelude::*;
 
+use secrecy::{ExposeSecret, SecretString};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
+use zeroize::{Zeroize, ZeroizeOnDrop};
 
 #[serde_as]
 #[derive(
-    Clone, PartialEq, Eq, Hash, Serialize, Deserialize, derive_more::Display, derive_more::Debug,
+    Clone,
+    PartialEq,
+    Eq,
+    Hash,
+    Serialize,
+    Deserialize,
+    derive_more::Display,
+    derive_more::Debug,
+    Zeroize,
+    ZeroizeOnDrop,
 )]
 #[display("{}", hex::encode(&self.0))]
 #[debug("{}", hex::encode(&self.0))]
@@ -14,15 +25,15 @@ pub struct EncryptedAppPassword(#[serde_as(as = "serde_with::hex::Hex")] Vec<u8>
 
 impl EncryptedAppPassword {
     pub fn new_by_deriving_and_encrypting(
-        app_password: String,
-        encryption_password: String,
+        app_password: SecretString,
+        encryption_password: SecretString,
     ) -> Self {
-        let encryption_key = PbHkdfSha256::derive_key(encryption_password);
+        let encryption_key = PbHkdfSha256::derive_key_from(encryption_password);
         Self::new_by_encrypting(app_password, encryption_key)
     }
 
-    pub fn new_by_encrypting(app_password: String, encryption_key: EncryptionKey) -> Self {
-        let sealed_box = AesGcm256::seal(app_password.as_bytes(), encryption_key);
+    pub fn new_by_encrypting(app_password: SecretString, encryption_key: EncryptionKey) -> Self {
+        let sealed_box = AesGcm256::seal(app_password.expose_secret().as_bytes(), encryption_key);
         let combined = sealed_box.combined();
         Self(combined)
     }
@@ -40,16 +51,16 @@ mod tests {
 
     #[test]
     fn test_encrypted_app_password() {
-        let app_password = "my_secret_app_password".to_string();
-        let encryption_pwd = "open sesame".to_string();
+        let app_password = SecretString::from("my_secret_app_password");
+        let encryption_pwd = SecretString::from("open sesame");
         let encrypted = EncryptedAppPassword::new_by_deriving_and_encrypting(
             app_password.clone(),
             encryption_pwd.clone(),
         );
         let decrypted = encrypted
-            .decrypt(PbHkdfSha256::derive_key(encryption_pwd))
+            .decrypt(PbHkdfSha256::derive_key_from(encryption_pwd))
             .unwrap();
 
-        assert_eq!(decrypted, app_password);
+        assert_eq!(decrypted, app_password.expose_secret());
     }
 }
