+++ crates/core/src/logic/mod.rs
@@ -1,6 +1,7 @@
 mod calendar_logic;
 mod command;
 mod create_pdf;
+mod encryption;
 mod file_path_logic;
 mod prepare_data;
 mod read_write_data;
@@ -11,6 +12,7 @@ mod serde_to_typst;
 pub use calendar_logic::*;
 pub use command::*;
 pub use create_pdf::*;
+pub use encryption::*;
 pub use file_path_logic::*;
 pub use prepare_data::*;
 pub use read_write_data::*;
