+++ crates/core/src/models/data/submodels/email/mod.rs
@@ -1,25 +1,20 @@
-mod aes_gcm_256;
-mod aes_gcm_sealed_box;
 mod attachment;
-
 #[allow(clippy::module_inception)]
 mod email;
 mod email_account;
 mod email_address;
+mod email_atom_template;
 mod email_credentials;
 mod email_settings;
-mod encrypted_app_password;
-mod pb_hkdf;
+mod proto_email;
 mod smtp_server;
 
-pub use aes_gcm_256::*;
-pub use aes_gcm_sealed_box::*;
 pub use attachment::*;
 pub use email::*;
 pub use email_account::*;
 pub use email_address::*;
+pub use email_atom_template::*;
 pub use email_credentials::*;
 pub use email_settings::*;
-pub use encrypted_app_password::*;
-pub use pb_hkdf::*;
+pub use proto_email::*;
 pub use smtp_server::*;
